+++ CHANGES.rst
@@ -81,6 +81,8 @@ Improvements
 - Add a "Picture" personal data field to registrations. When used, it allows including the
   picture provided by the user on badges/tickets (:pr:`6160`, thanks :user:`vtran99`)
 - Support ``~~text~~`` to strike-out text in markdown (:pr:`6166`)
+- Add experimental support for creating Google Wallet tickets (opt-in via :data:`ENABLE_GOOGLE_WALLET`
+  ``indico.conf`` setting) (:pr:`6028`, thanks :user:`openprojects`)
 
 Bugfixes
 ^^^^^^^^